pyproject.toml
[project]
name = "shortcat"
version = "0.1.0"
description = ""
authors = [
    {name = "user",email = "69367859+stefanodvx@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django (>=5.2.8,<6.0.0)",
    "python-dotenv (==1.2.1)",
    "djangorestframework (==3.16.1)",
    "django-cors-headers (==4.9.0)",
    "drf-yasg (==1.21.11)",
    "dj-rest-auth (==7.0.1)",
    "django-allauth (==65.13.1)",
    "requests (==2.32.5)",
    "drf-nested-routers (>=0.95.0,<0.96.0)",
    "whitenoise (>=6.11.0,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
